Delete useless test + rename some tests to represent the interceptor method
